integrate form into message component
