+++ apps/thundermoon_web/lib/thundermoon_web/live/chat_live/index.ex
@@ -10,8 +10,9 @@ defmodule ThundermoonWeb.ChatLive.Index do
 
   alias ThundermoonWeb.Presence
 
-  alias ThundermoonWeb.ChatLive.{Message, Form}
+  alias ThundermoonWeb.ChatLive.Message
 
+  @impl true
   def mount(_params, session, socket) do
     user = Accounts.get_user(session["current_user_id"])
     if connected?(socket), do: subscribe(user)
@@ -29,17 +30,30 @@ defmodule ThundermoonWeb.ChatLive.Index do
     {:noreply, socket}
   end
 
+  # this is triggered by the live_view event phx-submit
+  @impl true
+  def handle_event("send", %{"message" => %{"text" => text}}, socket) do
+    user = socket.assigns.current_user
+    message = %{user: user.username, text: text, user_id: user.id}
+    ChatMessages.add(message)
+    PubSub.broadcast(ThundermoonWeb.PubSub, "chat", {:send, message})
+    {:noreply, assign(socket, %{version: System.unique_integer()})}
+  end
+
   # this is triggered by the pubsub broadcast event
+  @impl true
   def handle_info({:send, message}, socket) do
     messages = [message | socket.assigns.messages]
     {:noreply, assign(socket, %{messages: messages})}
   end
 
+  @impl true
   def handle_info(:clear, socket) do
     {:noreply, assign(socket, %{messages: []})}
   end
 
   # this is triggered if someone leaves or joins the chat
+  @impl true
   def handle_info(
         %{
           event: "presence_diff",
